Move logic for migrating to hierarchical file structure to usecases/ for better reusability in backup/
